Specify base dir only in prod
Due to a Tauri issue, the base directory for the app (on macOS) is different in development (src-tauri) vs build (/). This change sets the correct dir based on the NODE_ENV set by Parcel, since `tauri dev` makes sense to use with `yarn start`. This won't work if `tauri dev` is run with assets created by `yarn build`.